tif_dirread.c: Logical cast working around compiler warning; tif_read.c: Correction of some error flags and parameter lists

+++ libtiff/tif_dirwrite.c
@@ -781,7 +781,7 @@ TIFFWriteDirectorySec(TIFF* tif, int isimage, int imagedone, uint64* pdiroff)
 			n+=4;
 			o++;
 		}
-		*(uint32*)n = tif->tif_nextdiroff;
+		*(uint32*)n = (uint32)tif->tif_nextdiroff;
 	}
 	else
 	{
